rename objects submod
