+++ src/entity/ai.rs
@@ -8,8 +8,8 @@ use crate::entity::act::{Action, Target, TargetCategory};
 use crate::entity::act::{InjectRnaVirus, Move, Pass, ProduceVirion};
 use crate::entity::control::{Ai, Controller};
 use crate::entity::object::Object;
-use crate::game::game_objects::GameObjects;
 use crate::game::game_state::GameState;
+use crate::game::objects::GameObjects;
 use crate::util::rng::RngExtended;
 use rand::seq::{IteratorRandom, SliceRandom};
 use serde::{Deserialize, Serialize};
