+++ src/entity/control.rs
@@ -1,6 +1,6 @@
 use crate::entity::act::{self, Action};
 use crate::entity::object::Object;
-use crate::game::game_objects::GameObjects;
+use crate::game::objects::GameObjects;
 use crate::game::game_state::GameState;
 #[cfg(not(target_arch = "wasm32"))]
 use serde::{Deserialize, Serialize};
