+++ src/ui/dialog/character.rs
@@ -1,4 +1,4 @@
-use crate::game::game_objects::GameObjects;
+use crate::game::objects::GameObjects;
 use crate::game::game_state::GameState;
 use crate::ui::dialog::InfoBox;
