+++ src/ui/menu/choose_action_menu.rs
@@ -1,5 +1,5 @@
-use crate::game::game_objects::GameObjects;
 use crate::game::game_state::GameState;
+use crate::game::objects::GameObjects;
 use crate::game::RunState;
 use crate::ui::menu::{Menu, MenuItem};
